+++ crates/validator/src/channel.rs
@@ -12,6 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use alloy::primitives::FixedBytes;
+use kailua_common::precondition::PreconditionValidationData;
+use risc0_zkvm::Receipt;
 use tokio::sync::mpsc::{channel, Receiver, Sender};
 
 /// A channel for two-way communication
@@ -38,4 +41,20 @@ impl<T> DuplexChannel<T> {
         };
         (channel_0, channel_1)
     }
+}
+
+#[derive(Clone, Debug)]
+#[allow(clippy::large_enum_variant)]
+pub enum Message {
+    // The proposal and its parent
+    Proposal {
+        index: u64,
+        precondition_validation_data: Option<PreconditionValidationData>,
+        l1_head: FixedBytes<32>,
+        agreed_l2_head_hash: FixedBytes<32>,
+        agreed_l2_output_root: FixedBytes<32>,
+        claimed_l2_block_number: u64,
+        claimed_l2_output_root: FixedBytes<32>,
+    },
+    Proof(u64, Option<Receipt>),
 }
